Revert ClientType placement
