+++ crates/ibc/src/core/ics02_client/mod.rs
@@ -1,6 +1,7 @@
 //! ICS 02: Client implementation for verifying remote IBC-enabled chains.
 
 pub mod client_state;
+pub mod client_type;
 pub mod consensus_state;
 pub mod error;
 pub mod events;
