+++ crates/ibc/src/core/ics24_host/identifier.rs
@@ -6,6 +6,7 @@ use derive_more::Into;
 
 use super::validate::*;
 use crate::clients::ics07_tendermint::client_type as tm_client_type;
+use crate::core::ics02_client::client_type::ClientType;
 use crate::core::ics24_host::validate::validate_client_identifier;
 
 use crate::core::ics24_host::error::ValidationError;
@@ -191,50 +192,6 @@ impl From<String> for ChainId {
     derive(borsh::BorshSerialize, borsh::BorshDeserialize)
 )]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-/// Type of the client, depending on the specific consensus algorithm.
-#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct ClientType(String);
-
-impl ClientType {
-    /// Constructs a new `ClientType` from the given `String` if it ends with a valid client identifier.
-    pub fn new(s: String) -> Result<Self, ValidationError> {
-        let s_trim = s.trim();
-        validate_client_type(s_trim)?;
-        Ok(Self(s_trim.to_string()))
-    }
-
-    /// Yields this identifier as a borrowed `&str`
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-}
-
-impl From<String> for ClientType {
-    /// Constructs a new `ClientType` from the given `String` without performing any validation.
-    fn from(value: String) -> Self {
-        Self(value)
-    }
-}
-
-impl Display for ClientType {
-    fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), FmtError> {
-        write!(f, "ClientType({})", self.0)
-    }
-}
-
-#[cfg_attr(
-    feature = "parity-scale-codec",
-    derive(
-        parity_scale_codec::Encode,
-        parity_scale_codec::Decode,
-        scale_info::TypeInfo
-    )
-)]
-#[cfg_attr(
-    feature = "borsh",
-    derive(borsh::BorshSerialize, borsh::BorshDeserialize)
-)]
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Into)]
 pub struct ClientId(String);
 
@@ -245,7 +202,7 @@ impl ClientId {
     ///
     /// ```
     /// # use ibc::core::ics24_host::identifier::ClientId;
-    /// # use ibc::core::ics24_host::identifier::ClientType;
+    /// # use ibc::core::ics02_client::client_type::ClientType;
     /// let tm_client_id = ClientId::new(ClientType::new("07-tendermint".to_string()).unwrap(), 0);
     /// assert!(tm_client_id.is_ok());
     /// tm_client_id.map(|id| { assert_eq!(&id, "07-tendermint-0") });
