+++ crates/ibc/src/mock/client_state.rs
@@ -1,3 +1,4 @@
+use crate::core::ics02_client::client_type::ClientType;
 use crate::core::ics24_host::path::{ClientConsensusStatePath, ClientStatePath};
 use crate::prelude::*;
 
@@ -15,7 +16,7 @@ use crate::core::ics02_client::error::ClientError;
 use crate::core::ics23_commitment::commitment::{
     CommitmentPrefix, CommitmentProofBytes, CommitmentRoot,
 };
-use crate::core::ics24_host::identifier::{ChainId, ClientId, ClientType};
+use crate::core::ics24_host::identifier::{ChainId, ClientId};
 use crate::core::ics24_host::Path;
 use crate::mock::client_state::client_type as mock_client_type;
 use crate::mock::consensus_state::MockConsensusState;
